Refactor navlog-viewer to use nanogui
This fixes opengl context errors with the old implementation.
