--- apps/navlog-viewer/imgs/main_icon.xpm
@@ -1,182 +1,0 @@
-/* XPM */
-static const char * main_icon_xpm[] = {
-"32 32 147 2",
-"  	c None",
-". 	c #559752",
-"+ 	c #50964E",
-"@ 	c #95C292",
-"# 	c #99C594",
-"$ 	c #67A761",
-"% 	c #3D8937",
-"& 	c #3E893C",
-"* 	c #90BF8B",
-"= 	c #8DC085",
-"- 	c #80BB75",
-"; 	c #6BB05E",
-"> 	c #429337",
-", 	c #3A8737",
-"' 	c #86BB7D",
-") 	c #82BC76",
-"! 	c #75B766",
-"~ 	c #6AB458",
-"{ 	c #61B14C",
-"] 	c #449E30",
-"^ 	c #2A871A",
-"/ 	c #368633",
-"( 	c #7CB872",
-"_ 	c #79B96A",
-": 	c #6DB55A",
-"< 	c #62B24B",
-"[ 	c #58B03D",
-"} 	c #51AF31",
-"| 	c #47AA26",
-"1 	c #309618",
-"2 	c #34862E",
-"3 	c #76B768",
-"4 	c #72B960",
-"5 	c #66B54F",
-"6 	c #5BB33F",
-"7 	c #52B131",
-"8 	c #4DB228",
-"9 	c #4EB426",
-"0 	c #52B828",
-"a 	c #41AA1F",
-"b 	c #2D9415",
-"c 	c #30882B",
-"d 	c #71B761",
-"e 	c #6DB957",
-"f 	c #61B646",
-"g 	c #57B436",
-"h 	c #50B42A",
-"i 	c #57BD2A",
-"j 	c #5BC12C",
-"k 	c #5EC42E",
-"l 	c #5AC12C",
-"m 	c #3FA81F",
-"n 	c #308A29",
-"o 	c #6EB75B",
-"p 	c #6BB951",
-"q 	c #5FB83F",
-"r 	c #56B72F",
-"s 	c #54BA29",
-"t 	c #59BF2B",
-"u 	c #62C830",
-"v 	c #65CB32",
-"w 	c #67CD34",
-"x 	c #66CD34",
-"y 	c #339619",
-"z 	c #308D27",
-"A 	c #6EB958",
-"B 	c #6BBC4D",
-"C 	c #5FBB3B",
-"D 	c #58BB2D",
-"E 	c #61C72F",
-"F 	c #68CE35",
-"G 	c #6BD138",
-"H 	c #6DD33A",
-"I 	c #66CD36",
-"J 	c #2C8517",
-"K 	c #308F27",
-"L 	c #6EBB57",
-"M 	c #6CBF4C",
-"N 	c #61BF3A",
-"O 	c #5CC12E",
-"P 	c #66CC33",
-"Q 	c #6AD037",
-"R 	c #6ED43B",
-"S 	c #71D73E",
-"T 	c #5BC531",
-"U 	c #348C1C",
-"V 	c #309227",
-"W 	c #70BE58",
-"X 	c #6FC24D",
-"Y 	c #65C33B",
-"Z 	c #61C630",
-"` 	c #6FD53C",
-" .	c #6DD43C",
-"..	c #4AAF29",
-"+.	c #1F6C11",
-"@.	c #339528",
-"#.	c #74C15A",
-"$.	c #73C650",
-"%.	c #6AC73F",
-"&.	c #65CA34",
-"*.	c #69CF36",
-"=.	c #5AC631",
-"-.	c #37901E",
-";.	c #074703",
-">.	c #34992A",
-",.	c #79C55F",
-"'.	c #79CA56",
-").	c #71CB46",
-"!.	c #6BCD39",
-"~.	c #65CD35",
-"{.	c #44AC25",
-"].	c #134A0A",
-"^.	c #000900",
-"/.	c #000000",
-"(.	c #379B2C",
-"_.	c #80C866",
-":.	c #80CD5E",
-"<.	c #78CF4E",
-"[.	c #59BF33",
-"}.	c #277214",
-"|.	c #022201",
-"1.	c #D22A2A",
-"2.	c #EA0000",
-"3.	c #BA0000",
-"4.	c #FF3333",
-"5.	c #EE0000",
-"6.	c #6FC059",
-"7.	c #7ECC60",
-"8.	c #499C32",
-"9.	c #14490C",
-"0.	c #AE2C22",
-"a.	c #DC0000",
-"b.	c #999833",
-"c.	c #1C5216",
-"d.	c #225D19",
-"e.	c #853C1C",
-"f.	c #C40000",
-"g.	c #777777",
-"h.	c #EF312F",
-"i.	c #D00100",
-"j.	c #E00000",
-"k.	c #AAAAAA",
-"l.	c #444444",
-"m.	c #DDDDDD",
-"n.	c #888888",
-"o.	c #222222",
-"p.	c #555555",
-"            . +                                                 ",
-"            @ # $ %                                             ",
-"          & * = - ; >                                           ",
-"          , ' ) ! ~ { ] ^                                       ",
-"          / ( _ : < [ } | 1                                     ",
-"          2 3 4 5 6 7 8 9 0 a b                                 ",
-"          c d e f g h 0 i j k l m                               ",
-"          n o p q r s t k u v w x y                             ",
-"          z A B C D j E v F G H I J                             ",
-"          K L M N O E P Q R S T U                               ",
-"          V W X Y Z P Q `  ...+.                                ",
-"          @.#.$.%.&.*.R =.-.;.                                  ",
-"          >.,.'.).!.~.{.].^././.          /./.                  ",
-"          (._.:.<.[.}.|.1.2.3./././.  /./.4.5./.                ",
-"            6.7.8.9.0.a.5.5./.b.b././.4.5./.5./.                ",
-"            c.d.e.f.2././.3./././.4.5././.g./.5./.              ",
-"        /./.h.i.j.5./.b.b././.4.5././.k.k.g./.5./.              ",
-"      /.4.4.4.5.5././.b.b././.5./.l.g.k.k.k.g./.5./.            ",
-"    /.4.m.5.5.4.4././.b.b./././.5./.l.g.g././.5.5.5././.    l.  ",
-"  /./.5.5.m.n.5.5.4.4././.b././.5./.l././.5.5.5.5.5.5.5././.  l.",
-"/.g./.5.3.3.5.m.n.5.5.4.4././.n./.5./.5.5.5.5.5.5.5.4.5.5.5./.  ",
-"/.l./.3././.3.5.5.n.n.5.5.4.5././.5.5.5.5.5.5.5.m.5.5.5.m.o./.l.",
-"  /./././././.3.5.5.5.n.n.5.5.5.5.5.5.5.5.5.5.5.5.n.p.3.o.p./.  ",
-"      /.g./././.3.3.5.5.5.n.p.5.5.5.5.5.4.4.5.5.5.3.p.3.p.n./.l.",
-"      /.n./././././.3.3.3.3.3.n.p.5.5.5.5.5.m.o.3.3.3.3./././.  ",
-"      /./././././.  /././././.3.5.n.p.5.5.5.o.p.3.3./././././.  ",
-"        /./././.        /./././.3.5.5.5.5.5.p.n././././././.    ",
-"                        /.g./././.3.3.5.5./././././././.        ",
-"                        /.n./././././.3.3./././././.            ",
-"                        /./././././.  /././././.                ",
-"                          /./././.                              ",
-"                                                                "};
