--- apps/navlog-viewer/navlog-viewer_GUI_design.cbp
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<CodeBlocks_project_file>
-	<FileVersion major="1" minor="6" />
-	<Project>
-		<Option title="navlog-viewer_GUI_design" />
-		<Option pch_mode="2" />
-		<Option compiler="gcc" />
-		<Build>
-			<Target title="Debug">
-				<Option output="bin\Debug\navlog-viewer_GUI_design" prefix_auto="1" extension_auto="1" />
-				<Option object_output="obj\Debug\" />
-				<Option type="0" />
-				<Option compiler="gcc" />
-				<Option projectLinkerOptionsRelation="2" />
-				<Compiler>
-					<Add option="-g" />
-				</Compiler>
-			</Target>
-			<Target title="Release">
-				<Option output="bin\Release\navlog-viewer_GUI_design" prefix_auto="1" extension_auto="1" />
-				<Option object_output="obj\Release\" />
-				<Option type="0" />
-				<Option compiler="gcc" />
-				<Option projectLinkerOptionsRelation="2" />
-				<Compiler>
-					<Add option="-O2" />
-				</Compiler>
-				<Linker>
-					<Add option="-s" />
-				</Linker>
-			</Target>
-		</Build>
-		<Compiler>
-			<Add option="`wx-config --cflags`" />
-			<Add option="-Wall" />
-		</Compiler>
-		<Linker>
-			<Add option="`wx-config --libs`" />
-		</Linker>
-		<Unit filename="navlog_viewer_GUI_designApp.cpp" />
-		<Unit filename="navlog_viewer_GUI_designApp.h" />
-		<Unit filename="navlog_viewer_GUI_designMain.cpp" />
-		<Unit filename="navlog_viewer_GUI_designMain.h" />
-		<Unit filename="wxsmith\navlog_viewer_GUI_designdialog.wxs" />
-		<Extensions>
-			<envvars />
-			<code_completion />
-			<wxsmith version="1">
-				<gui name="wxWidgets" src="navlog_viewer_GUI_designApp.cpp" main="navlog_viewer_GUI_designDialog" init_handlers="necessary" language="CPP" />
-				<resources>
-					<wxDialog wxs="wxsmith/navlog_viewer_GUI_designdialog.wxs" src="navlog_viewer_GUI_designMain.cpp" hdr="navlog_viewer_GUI_designMain.h" name="navlog_viewer_GUI_designDialog" language="CPP" />
-				</resources>
-			</wxsmith>
-			<debugger />
-		</Extensions>
-	</Project>
-</CodeBlocks_project_file>
