--- apps/navlog-viewer/resource.rc
@@ -1,3 +1,0 @@
-aaaa ICON "imgs/main_icon.ico"
-
-#include "wx/msw/wx.rc"
